refactoring: extractFileName, extractFileType 이동
