+++ src/main/java/org/etmetmy/bn_server/domain/file/dto/request/FileCreateRequest.java
@@ -7,9 +7,6 @@ import lombok.NoArgsConstructor;
 import org.etmetmy.bn_server.domain.file.entity.File;
 import org.etmetmy.bn_server.domain.post.entity.Post;
 
-import static org.etmetmy.bn_server.domain.file.service.FileServiceImpl.extractFileName;
-import static org.etmetmy.bn_server.domain.file.service.FileServiceImpl.extractFileType;
-
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
@@ -20,9 +17,8 @@ public class FileCreateRequest {
 
 
     public static class Converter {
-        /**
-         * File 엔티티 생성
-         */
+
+        // File 엔티티 생성
         public static File toEntity(String fileUrl, Long fileSize, Post post, Long uploadedBy) {
             return File.builder()
                     .post(post)
@@ -33,6 +29,18 @@ public class FileCreateRequest {
                     .uploadedBy(uploadedBy)
                     .isDeleted(false)
                     .build();
+        }
+
+        // URL 에서 파일명 추출
+        public static String extractFileName(String url) {
+            int lastSlash = url.lastIndexOf('/');
+            return lastSlash >= 0 ? url.substring(lastSlash + 1) : "unknown";
+        }
+
+        // URL 에서 파일 확장자 추출
+        public static String extractFileType(String url) {
+            int lastDot = url.lastIndexOf('.');
+            return lastDot >= 0 ? url.substring(lastDot + 1).toLowerCase() : "unknown";
         }
     }
 }
