+++ src/main/java/org/etmetmy/bn_server/domain/post/dto/request/PostCreateRequest.java
@@ -86,4 +86,5 @@ public class PostCreateRequest {
                     .build();
         }
     }
+
 }
